Move rest of core_test into specific unit tests.
