+++ src/test/lib/template-factory_test.ts
@@ -1,0 +1,26 @@
+/**
+ * @license
+ * Copyright (c) 2017 The Polymer Project Authors. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * The complete set of authors may be found at
+ * http://polymer.github.io/AUTHORS.txt
+ * The complete set of contributors may be found at
+ * http://polymer.github.io/CONTRIBUTORS.txt
+ * Code distributed by Google as part of the polymer project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
+
+import {html, templateFactory} from '../../index.js';
+
+const assert = chai.assert;
+
+suite('TemplateFactory', () => {
+  test(
+      'returns identical templates for TemplateResults for same strings',
+      () => {
+        const t = () => html``;
+        assert.strictEqual(templateFactory(t()), templateFactory(t()));
+      });
+});
