+++ test/index.html
@@ -5,7 +5,10 @@
     <script src="../node_modules/@webcomponents/webcomponentsjs/webcomponents-bundle.js"></script>
   </head>
   <body>
-    <script type="module" src="./core_test.js"></script>
+    <script type="module" src="./index_test.js"></script>
+    <script type="module" src="./lib/template-result_test.js"></script>
+    <script type="module" src="./lib/template_test.js"></script>
+    <script type="module" src="./lib/template-factory_test.js"></script>
     <script type="module" src="./lib/render_test.js"></script>
     <script type="module" src="./lib/parts_test.js"></script>
     <script type="module" src="./lib/lit-extended_test.js"></script>
